:recycle: Move response related function as `response.go`
